+++ website/src/com/project/website/shared/client/html5/Selection.java
@@ -5,23 +5,23 @@ import com.google.gwt.dom.client.Node;
 /* Selection IDL from http://html5.org/specs/dom-range.html#selection
  * interface Selection {
       readonly attribute Node? anchorNode;
-      readonly attribute unsigned long anchorOffset;
+      readonly attribute unsigned int anchorOffset;
       readonly attribute Node? focusNode;
-      readonly attribute unsigned long focusOffset;
+      readonly attribute unsigned int focusOffset;
 
       readonly attribute boolean isCollapsed;
-      void               collapse(Node parentNode, unsigned long offset);
+      void               collapse(Node parentNode, unsigned int offset);
       void               collapseToStart();
       void               collapseToEnd();
 
-      void               extend(Node parentNode, unsigned long offset);
+      void               extend(Node parentNode, unsigned int offset);
       void               modify(DOMString alter, DOMString direction, DOMString granularity);
 
       void               selectAllChildren(Node parentNode);
       void               deleteFromDocument();
 
-      readonly attribute unsigned long rangeCount;
-      Range              getRangeAt(unsigned long index);
+      readonly attribute unsigned int rangeCount;
+      Range              getRangeAt(unsigned int index);
       void               addRange(Range range);
       void               removeRange(Range range);
       void               removeAllRanges();
@@ -31,24 +31,24 @@ import com.google.gwt.dom.client.Node;
  */
 public interface Selection  {
     Node getAnchorNode();
-    long getAnchorOffset();
+    int getAnchorOffset();
     Node getFocusNode();
-    long getFocusOffset();
+    int getFocusOffset();
 
     boolean isCollapsed();
 
-    void collapse(Node parentNode, long offset);
+    void collapse(Node parentNode, int offset);
     void collapseToStart();
     void collapseToEnd();
 
-    void extend(Node parentNode, long offset);
+    void extend(Node parentNode, int offset);
     void modify(String alter, String direction, String granularity);
 
     void selectAllChildren(Node parentNode);
     void deleteFromDocument();
 
-    long getRangeCount();
-    Range getRangeAt(long index);
+    int getRangeCount();
+    Range getRangeAt(int index);
     void addRange(Range range);
     void removeRange(Range range);
     void removeAllRanges();
